Make Constants a class and move in seperate files
Makes init/teardown simpler (and avoids even more global state)

Header cleanup
