+++ src/exe_reader.cpp
@@ -16,16 +16,17 @@
  */
 
 // All of this code is unused on EMSCRIPTEN. *Do not use it*!
+
 #ifndef EMSCRIPTEN
 
 #include "exe_reader.h"
 #include "image_bmp.h"
 #include "output.h"
-#include "utils.h"
+#include "player.h"
 #include <algorithm>
 #include <iostream>
-#include <fstream>
 #include <zlib.h>
+#include <unordered_map>
 
 namespace {
 	// hashes of known RPG_RT startup logos
@@ -565,13 +566,13 @@ int EXEReader::FileInfo::GetEngineType(int& mp_version) const {
 	return Player::EngineNone;
 }
 
-std::map<Player::GameConstantType, int32_t> EXEReader::GetOverriddenGameConstants() {
-	std::map<Player::GameConstantType, int32_t> game_constants;
+std::unordered_map<Game_Constants::ConstantType, int32_t> EXEReader::GetOverriddenGameConstants() {
+	std::unordered_map<Game_Constants::ConstantType, int32_t> game_constants;
 
-	auto apply_known_config = [&](Player::Constants::KnownPatchConfigurations conf) {
-		Output::Debug("Assuming known patch config '{}'", Player::Constants::kKnownPatchConfigurations.tag(static_cast<int>(conf)));
-		auto it_conf = Player::Constants::known_patch_configurations.find(conf);
-		assert(it_conf != Player::Constants::known_patch_configurations.end());
+	auto apply_known_config = [&](Game_Constants::KnownPatchConfigurations conf) {
+		Output::Debug("Assuming known patch config '{}'", Game_Constants::kKnownPatchConfigurations.tag(static_cast<int>(conf)));
+		auto it_conf = Game_Constants::known_patch_configurations.find(conf);
+		assert(it_conf != Game_Constants::known_patch_configurations.end());
 
 		for (auto it = it_conf->second.begin(); it != it_conf->second.end(); ++it) {
 			game_constants[it->first] = it->second;
@@ -589,7 +590,7 @@ std::map<Player::GameConstantType, int32_t> EXEReader::GetOverriddenGameConstant
 	switch (file_info.code_size) {
 		case 0x9CC00: // RM2K 1.62
 			if (check_for_string(file_info.code_ofs + 0x07DAA6, "XXX" /* 3x "POP EAX" */)) {
-				apply_known_config(Player::Constants::KnownPatchConfigurations::StatDelimiter);
+				apply_known_config(Game_Constants::KnownPatchConfigurations::StatDelimiter);
 			}
 			break;
 		case 0xC8E00: // RM2K3 1.0.8.0
@@ -597,15 +598,15 @@ std::map<Player::GameConstantType, int32_t> EXEReader::GetOverriddenGameConstant
 			// to translate this string in RPG_RT. So this segment can be used to reliably detect
 			// the patch without having to read all individual constant values from the EXE
 			if (check_for_string(file_info.code_ofs + 0x08EBE0, "NoTitolo")) {
-				apply_known_config(Player::Constants::KnownPatchConfigurations::Rm2k3_Italian_WD_108);
+				apply_known_config(Game_Constants::KnownPatchConfigurations::Rm2k3_Italian_WD_108);
 			}
 			if (check_for_string(file_info.code_ofs + 0x09D279, "XXX" /* 3x "POP EAX" */)) {
-				apply_known_config(Player::Constants::KnownPatchConfigurations::StatDelimiter);
+				apply_known_config(Game_Constants::KnownPatchConfigurations::StatDelimiter);
 			}
 			break;
 		case 0xC9000: // RM2K3 1.0.9.1
 			if (check_for_string(file_info.code_ofs + 0x09C5AD, "XXX" /* 3x "POP EAX" */)) {
-				apply_known_config(Player::Constants::KnownPatchConfigurations::StatDelimiter);
+				apply_known_config(Game_Constants::KnownPatchConfigurations::StatDelimiter);
 			}
 			break;
 	}
