+++ src/main_data.cpp
@@ -19,6 +19,7 @@
 #include <cstdlib>
 #include "main_data.h"
 #include "filefinder_rtp.h"
+#include "game_constants.h"
 #include "game_destiny.h"
 #include "game_system.h"
 #include "game_actors.h"
@@ -70,6 +71,7 @@ namespace Main_Data {
 	std::unique_ptr<Game_DynRpg> game_dynrpg;
 	std::unique_ptr<Game_Ineluki> game_ineluki;
 	std::unique_ptr<Game_Destiny> game_destiny;
+	std::unique_ptr<Game_Constants> game_constants;
 	std::unique_ptr<Game_Switches> game_switches_global;
 	std::unique_ptr<Game_Variables> game_variables_global;
 
@@ -112,6 +114,7 @@ void Main_Data::Init() {
 void Main_Data::Cleanup() {
 	Game_Map::Quit();
 
+	game_constants.reset();
 	game_switches.reset();
 	game_screen.reset();
 	game_pictures.reset();
