+++ src/player.h
@@ -59,36 +59,6 @@ namespace Player {
 		PatchOverride = 1 << 16
 	};
 
-	enum class GameConstantType {
-		MinVarLimit,
-		MaxVarLimit,
-		MaxActorHP,
-		MaxActorSP,
-		MaxStatBaseValue,
-		MaxStatBattleValue,
-		MaxDamageValue,
-		MaxExpValue,
-		MaxLevel,
-		MaxGoldValue,
-		MaxItemCount,
-		MaxSaveFiles
-	};
-
-	static constexpr auto kGameConstantType = lcf::makeEnumTags<GameConstantType>(
-		"MinVarLimit",
-		"MaxVarLimit",
-		"MaxActorHP",
-		"MaxActorSP",
-		"MaxStatBaseValue",
-		"MaxStatBattleValue",
-		"MaxDamageValue",
-		"MaxExpValue",
-		"MaxLevel",
-		"MaxGoldValue",
-		"MaxItemCount",
-		"MaxSaveFiles"
-	);
-
 	/**
 	 * Initializes EasyRPG Player.
 	 *
@@ -473,71 +443,6 @@ namespace Player {
 	std::optional<bool> GetRuntimeFlag(Game_Interpreter_Shared::StateRuntimeFlagRef field_on, Game_Interpreter_Shared::StateRuntimeFlagRef field_off);
 #endif
 
-	namespace Constants {
-		using Var_t = int32_t;
-
-		static constexpr int32_t max_level_2k = 50;
-		static constexpr int32_t max_level_2k3 = 99;
-
-		void GetVariableLimits(Var_t& min_var, Var_t& max_var);
-
-		int32_t MaxActorHpValue();
-		int32_t MaxActorSpValue();
-
-		int32_t MaxEnemyHpValue();
-		int32_t MaxEnemySpValue();
-
-		int32_t MaxStatBaseValue();
-		int32_t MaxStatBattleValue();
-		int32_t MaxDamageValue();
-
-		int32_t MaxExpValue();
-		int32_t MaxLevel();
-
-		int32_t MaxGoldValue();
-		int32_t MaxItemCount();
-		int32_t MaxSaveFiles();
-
-		extern std::map<GameConstantType, int32_t> constant_overrides;
-
-		bool TryGetOverriddenConstant(GameConstantType const_type, int32_t& out_value);
-		void OverrideGameConstant(GameConstantType const_type, int32_t value);
-		void ResetOverrides();
-		void PrintActiveOverrides();
-
-		enum class KnownPatchConfigurations {
-			Rm2k3_Italian_WD_108,		// Italian "WhiteDragon" patch
-			StatDelimiter,
-			LAST
-		};
-		static constexpr auto kKnownPatchConfigurations = lcf::makeEnumTags<KnownPatchConfigurations>(
-			"Rm2k3 Italian 1.08",
-			"StatDelimiter"
-		);
-
-		static_assert(kKnownPatchConfigurations.size() == static_cast<size_t>(KnownPatchConfigurations::LAST));
-
-		using T = Player::GameConstantType;
-		using patch_config = std::map<Player::GameConstantType, int32_t>;
-		inline const std::map<KnownPatchConfigurations, patch_config> known_patch_configurations = {
-			{
-				KnownPatchConfigurations::Rm2k3_Italian_WD_108, {
-					{ T::MinVarLimit,	-999999999 },
-					{ T::MaxVarLimit,	 999999999 },
-					{ T::MaxActorHP,	     99999 },
-					{ T::MaxActorSP,          9999 },
-					{ T::MaxStatBaseValue,    9999 },
-					{ T::MaxDamageValue,     99999 },
-					{ T::MaxGoldValue,     9999999 }
-			}},{
-			KnownPatchConfigurations::StatDelimiter, {
-				{ T::MaxActorHP,           9999999 },
-				{ T::MaxActorSP,           9999999 },
-				{ T::MaxStatBaseValue,      999999 },
-				{ T::MaxStatBattleValue,    999999 }
-		}}
-		};
-	}
 }
 
 inline bool Player::IsRPG2k() {
