+++ tests/mock_game.cpp
@@ -1,6 +1,9 @@
 #include "mock_game.h"
 #include "game_actors.h"
+#include "game_constants.h"
 #include "game_system.h"
+#include "main_data.h"
+#include <memory>
 
 static lcf::rpg::Terrain MakeTerrain() {
 	return {};
@@ -22,6 +25,7 @@ MockGame::MockGame(MockMap maptag) {
 	lcf::Data::terrains.push_back(MakeTerrain());
 	lcf::Data::chipsets.push_back(MakeChipset());
 
+	Main_Data::game_constants = std::make_unique<Game_Constants>();
 	Main_Data::game_actors = std::make_unique<Game_Actors>();
 	Main_Data::game_party = std::make_unique<Game_Party>();
